% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhpc.R
\name{mhpc}
\alias{mhpc}
\title{Mining combined causes}
\usage{
mhpc(data, variant = "B", alpha = 0.05, supp = 0.05, cor.pvalue = 0.01)
}
\arguments{
\item{data:}{binary data; variant: B for MH-PC-B or F for MH-PC-F}

\item{alpha:}{threshold of the conditional independece test}

\item{supp:}{threshold of the pruning scheme}

\item{cor.pvalue:}{threshold of the pruning scheme}
}
\value{
all single and combined causes
}
\description{
MH-PC is implemented based on the semi-interleaved HITOn-PC in bnlearn
}
\author{
Saisai Ma
}
